feat: raise on inverted min max except longitude antemeridian crossing (#396)
