+++ copernicusmarine/__init__.py
@@ -29,7 +29,6 @@ from copernicusmarine.core_functions.exceptions import (
     CoordinatesOutOfDatasetBounds,
     DatasetUpdating,
     FormatNotSupported,
-    MinimumLongitudeGreaterThanMaximumLongitude,
     NetCDFCompressionNotAvailable,
     NoServiceAvailable,
     NotEnoughPlatformMetadata,
@@ -88,7 +87,6 @@ __all__ = [
     "FormatNotSupported",
     "GeographicalExtent",
     "InvalidUsernameOrPassword",
-    "MinimumLongitudeGreaterThanMaximumLongitude",
     "NetCDFCompressionNotAvailable",
     "NoServiceAvailable",
     "NotEnoughPlatformMetadata",
