+++ tests/test_sparse_subset.py
@@ -2,11 +2,9 @@ import pathlib
 from json import loads
 
 import pandas as pd
+import pytest
 
 from copernicusmarine import read_dataframe
-from copernicusmarine.core_functions.exceptions import (
-    MinimumLongitudeGreaterThanMaximumLongitude,
-)
 from copernicusmarine.download_functions.download_sparse import (
     COLUMNS_ORDER_DEPTH,
     COLUMNS_ORDER_ELEVATION,
@@ -160,7 +158,10 @@ class TestSparseSubset:
         assert list(df.columns) == COLUMNS_ORDER_ELEVATION
 
     def test_error_raises_for_inverted_longitude(self):
-        try:
+        with pytest.raises(
+            ValueError,
+            match="Minimum longitude greater than maximum longitude",
+        ):
             _ = read_dataframe(
                 dataset_id="cmems_obs-wave_glo_phy-swh_nrt_cfo-l3_PT1S",
                 variables=["VAVH", "VAVH_UNFILTERED", "WIND_SPEED"],
@@ -171,20 +172,15 @@ class TestSparseSubset:
                 start_datetime="01-01-2023",
                 end_datetime="02-01-2023",
             )
-        except MinimumLongitudeGreaterThanMaximumLongitude as e:
-            assert (
-                "--minimum-longitude option must be smaller "
-                "or equal to --maximum-longitude" in e.__str__()
-            )
 
     def test_if_ask_for_empty_dataframe_it_works(self, caplog):
         df = read_dataframe(
             dataset_id="cmems_obs-wave_glo_phy-swh_nrt_cfo-l3_PT1S",
             variables=["VAVH", "VAVH_UNFILTERED", "WIND_SPEED"],
-            maximum_latitude=30.1875,
-            minimum_latitude=30.188,
-            maximum_longitude=32,
+            minimum_latitude=30.1876,
+            maximum_latitude=30.1877,
             minimum_longitude=31.9,
+            maximum_longitude=32,
             start_datetime="01-01-2023",
             end_datetime="02-01-2023",
         )
